Reupload and overhauls
Reuploaded all files to new repo, completely overhauled the user interface for adding and managing member records, added aesthetic implementations for the GUI, overhauled the member search function to include validation, and added a new page for listing out all records associated with the search query. Removed some files related to member record management (such as dedicated editing and deleting pages) and merged them into a more clean and compact process, which are the aforementioned pages the overhauls are made in.